Refers issue #68
Move LogRuntime to core/api/runtime directory.
